+++ flang/include/flang/Lower/CustomIntrinsicCall.h
@@ -25,6 +25,7 @@
 #define FORTRAN_LOWER_CUSTOMINTRINSICCALL_H
 
 #include "flang/Lower/AbstractConverter.h"
+#include "flang/Optimizer/Builder/IntrinsicCall.h"
 #include <functional>
 #include <optional>
 
@@ -48,6 +49,8 @@ bool intrinsicRequiresCustomOptionalHandling(
 /// Type of callback to be provided to prepare the arguments fetching from an
 /// actual argument expression.
 using OperandPrepare = std::function<void(const Fortran::lower::SomeExpr &)>;
+using OperandPrepareAs = std::function<void(const Fortran::lower::SomeExpr &,
+                                            fir::LowerIntrinsicArgAs)>;
 
 /// Type of the callback to inquire about an argument presence, once the call
 /// preparation was done. An absent optional means the argument is statically
@@ -61,8 +64,9 @@ using OperandPresent = std::function<std::optional<mlir::Value>(std::size_t)>;
 /// verified. This means the getter callback can dereference the argument
 /// without any special care.
 /// For elemental intrinsics, the getter must provide the current iteration
-/// element value.
-using OperandGetter = std::function<fir::ExtendedValue(std::size_t)>;
+/// element value. If the boolean argument is true, the callback must load the
+/// argument before returning it.
+using OperandGetter = std::function<fir::ExtendedValue(std::size_t, bool)>;
 
 /// Given a callback \p prepareOptionalArgument to prepare optional
 /// arguments and a callback \p prepareOtherArgument to prepare non-optional
@@ -78,7 +82,7 @@ void prepareCustomIntrinsicArgument(
     const Fortran::evaluate::SpecificIntrinsic &intrinsic,
     std::optional<mlir::Type> retTy,
     const OperandPrepare &prepareOptionalArgument,
-    const OperandPrepare &prepareOtherArgument, AbstractConverter &converter);
+    const OperandPrepareAs &prepareOtherArgument, AbstractConverter &converter);
 
 /// Given a callback \p getOperand to generate a reference to the i-th argument,
 /// and a callback \p isPresentCheck to test if an argument is present, this