+++ libc/src/__support/str_to_integer.h
@@ -10,6 +10,7 @@
 #define LIBC_SRC_SUPPORT_STR_TO_INTEGER_H
 
 #include "src/__support/CPP/limits.h"
+#include "src/__support/CPP/type_traits.h"
 #include "src/__support/common.h"
 #include "src/__support/ctype_utils.h"
 #include "src/__support/str_to_num_result.h"
@@ -142,7 +143,9 @@ LIBC_INLINE StrToNumResult<T> strtointeger(const char *__restrict src,
       return {cpp::numeric_limits<T>::min(), str_len, error_val};
   }
 
-  return {is_positive ? static_cast<T>(result) : -static_cast<T>(result),
+  return {is_positive
+              ? static_cast<T>(result)
+              : static_cast<T>(-static_cast<cpp::make_unsigned_t<T>>(result)),
           str_len, error_val};
 }