+++ lld/MachO/Driver.cpp
@@ -1921,7 +1921,8 @@ bool macho::link(ArrayRef<const char *> argsArr, llvm::raw_ostream &stdoutOS,
     if (config->deadStrip)
       markLive();
 
-    objc::checkCategories();
+    if (args.hasArg(OPT_check_category_conflicts))
+      objc::checkCategories();
 
     // ICF assumes that all literals have been folded already, so we must run
     // foldIdenticalLiterals before foldIdenticalSections.